gabung fungsi dan perbaiki navigasi lagi

+++ LaundryGo/app/src/main/java/com/android/laundrygo/model/LaundryService.kt
@@ -6,7 +6,7 @@ data class LaundryService(
     @DocumentId val id: String = "", // Untuk menyimpan ID dokumen dari Firestore
     val title: String = "",
     val description: String = "",
-    val price: Long = 0,
+    val price: Double = 0.0,
     val unit: String = "",
     val category: String = ""
 )
